+++ PowerEditor/src/WinControls/DockingWnd/DockingCont.h
@@ -113,7 +113,7 @@ public:
 		return _isFloating;
 	}
 
-	INT getElementCnt() {
+	size_t getElementCnt() {
 		return _vTbData.size();
 	}
 
@@ -142,7 +142,7 @@ public:
 	};
 
     virtual void destroy() {
-		for (INT iTb = _vTbData.size(); iTb > 0; iTb--)
+		for (int iTb = (int)_vTbData.size(); iTb > 0; iTb--)
 		{
 			delete _vTbData[iTb-1];
 		}
